requirements.txt
langchain-openai>=0.0.2
google-generativeai>=0.3.2
langchain-google-genai>=0.0.3
python-dotenv>=1.0.0
pydantic>=2.5.2
pydantic-settings>=2.0.0
fastapi>=0.104.1
uvicorn>=0.24.0
websockets>=12.0
python-multipart>=0.0.6
requests>=2.31.0 